+++ extensions/plugin-history-sync/src/RoutesContext.tsx
@@ -2,13 +2,13 @@ import { createContext, useContext } from "react";
 
 import type { ActivityRoute } from "./ActivityRoute";
 
-export const RoutesContext = createContext<ActivityRoute[]>([]);
+export const RoutesContext = createContext<ActivityRoute<unknown>[]>([]);
 
-interface RoutesProviderProps {
-  routes: ActivityRoute[];
+interface RoutesProviderProps<T> {
+  routes: ActivityRoute<T>[];
   children: React.ReactNode;
 }
-export const RoutesProvider: React.FC<RoutesProviderProps> = (props) => (
+export const RoutesProvider = <T,>(props: RoutesProviderProps<T>) => (
   <RoutesContext.Provider value={props.routes}>
     {props.children}
   </RoutesContext.Provider>