+++ extensions/plugin-history-sync/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stackflow/plugin-history-sync
 
+## 1.4.0-canary.1
+
+### Patch Changes
+
+- fix(plugin-history-sync): decode search params
+
+## 1.4.0-canary.0
+
+### Minor Changes
+
+- 36613e35: Sort routes by variable count and refactor useRoutes(), normalizeRouteInput() function
+
+### Patch Changes
+
+- feat: add decode interface
+- Updated dependencies
+  - @stackflow/react@1.1.8-canary.0
+
 ## 1.3.18
 
 ### Patch Changes